lock logic for both buttons and table
